refactor: split engine/workflow/validation.py into focused modules
